[chore #163664493] Fix issues on ft #163388520
